+++ pyfakefs/fake_path.py
@@ -165,19 +165,7 @@ class FakePathModule:
 
     def isabs(self, path: AnyStr) -> bool:
         """Return `True` if path is an absolute pathname."""
-        empty = matching_string(path, "")
-        if self.filesystem.is_windows_fs:
-            drive, path = self.splitdrive(path)
-        else:
-            drive = empty
-        path = make_string_path(path)
-        if not self.filesystem.starts_with_sep(path):
-            return False
-        if self.filesystem.is_windows_fs and sys.version_info >= (3, 13):
-            # from Python 3.13 on, a path under Windows starting with a single separator
-            # (e.g. not a drive and not an UNC path) is no more considered absolute
-            return drive != empty
-        return True
+        return self.filesystem.isabs(path)
 
     def isdir(self, path: AnyStr) -> bool:
         """Determine if path identifies a directory."""
@@ -331,6 +319,7 @@ class FakePathModule:
     def relpath(self, path: AnyStr, start: AnyStr | None = None) -> AnyStr:
         """We mostly rely on the native implementation and adapt the
         path separator."""
+        path = self.os.fspath(path)
         if not path:
             raise ValueError("no path specified")
         path = make_string_path(path)
@@ -355,6 +344,7 @@ class FakePathModule:
         """Return the canonical path of the specified filename, eliminating any
         symbolic links encountered in the path.
         """
+        filename = self.os.fspath(filename)
         has_allow_missing = hasattr(os.path, "ALLOW_MISSING")
         if has_allow_missing and strict == os.path.ALLOW_MISSING:  # type: ignore[attr-defined]
             # ignores non-existing file, but not other errors
@@ -413,13 +403,32 @@ class FakePathModule:
         if self.filesystem.starts_with_root_path(rest):
             if self.filesystem.is_windows_fs:
                 drive, rest = self.filesystem.splitdrive(rest)
-                if not drive:
+                is_abs_path = self.filesystem.starts_with_sep(rest)
+                if drive and is_abs_path:
+                    # absolute path starting with drive letter
+                    path = drive + sep
+                else:
                     cwd = matching_string(path, self.filesystem.cwd)
-                    drive, _ = self.filesystem.splitdrive(cwd)
-                path = drive + sep
+                    cwd_drive, cwd_path = self.filesystem.splitdrive(cwd)
+                    if not drive:
+                        if is_abs_path:
+                            # absolute path without drive letter - relative to current drive
+                            path = cwd_drive + sep
+                        else:
+                            # relative path without drive letter
+                            path = cwd_drive + cwd_path + sep
+                    elif drive.lower() == cwd_drive.lower():
+                        # relative path with drive letter matching the current drive
+                        path = cwd_drive + cwd_path + sep
+                    else:
+                        # invalid relative path - do not change
+                        path = drive + rest
+                        rest = rest[:0]
+                if is_abs_path:
+                    rest = rest[1:]
             else:
                 path = sep
-            rest = rest[1:]
+                rest = rest[1:]
 
         while rest:
             name, _, rest = rest.partition(sep)